+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Features
 
 - Report start up crashes (#2220)
+- Add segment property to user (#2234)
 
 ## 7.26.0